AC-206: Extracted Fframework only sniffs to a separate namespace
