+++ .github/workflows/php.yml
@@ -59,4 +59,7 @@ jobs:
       run: vendor/bin/phpunit
 
     - name: Run code style suite
-      run: vendor/bin/phpcs --standard=Magento2 Magento2/Helpers/ Magento2/Sniffs
+      run: vendor/bin/phpcs --standard=Magento2 Magento2/Helpers Magento2/Sniffs Magento2Framework/Sniffs
+
+    - name: Run framework suite
+      run: vendor/bin/phpcs --standard=Magento2Framework Magento2/Helpers Magento2/Sniffs Magento2Framework/Sniffs
