+++ Magento2Framework/Tests/Header/CopyrightAnotherExtensionsFilesUnitTest.1.xml
