+++ Magento2Framework/Tests/Header/CopyrightAnotherExtensionsFilesUnitTest.2.js
