+++ Magento2Framework/Tests/Header/CopyrightAnotherExtensionsFilesUnitTest.3.xml
