+++ Magento2Framework/Tests/Header/CopyrightAnotherExtensionsFilesUnitTest.4.js
