+++ Magento2Framework/Tests/Header/CopyrightAnotherExtensionsFilesUnitTest.5.less
