+++ Magento2Framework/Tests/Header/CopyrightUnitTest.2.inc
