+++ Magento2Framework/Tests/Header/CopyrightUnitTest.5.inc
