+++ Magento2Framework/ruleset.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0"?>
-<ruleset name="Magento2 Only">
-    <description>Magento Coding Standard sniffs applicable for Magento Open Source development only</description>
-    <rule ref="Magento2.Legacy.License">
+<ruleset name="Magento2Framework">
+    <description>Magento Coding Standard sniffs applicable for the framework testing only</description>
+    <rule ref="Magento2Framework.Header.License">
         <severity>5</severity>
         <type>warning</type>
     </rule>
-    <rule ref="Magento2.Legacy.CopyrightAnotherExtensionsFiles">
+    <rule ref="Magento2Framework.Header.CopyrightAnotherExtensionsFiles">
         <severity>5</severity>
         <type>warning</type>
         <exclude-pattern>*\.php$</exclude-pattern>
         <exclude-pattern>*\.phtml$</exclude-pattern>
     </rule>
-    <rule ref="Magento2.Legacy.Copyright">
+    <rule ref="Magento2Framework.Header.Copyright">
         <severity>5</severity>
         <type>warning</type>
         <include-pattern>*\.php$</include-pattern>
